+++ test/test_base.cpp
@@ -4,9 +4,9 @@
 #include <string>
 #include <numeric>
 
-#include "tilt/codegen/loopgen.h"
-#include "tilt/codegen/llvmgen.h"
-#include "tilt/codegen/vinstr.h"
+#include "tilt/pass/codegen/loopgen.h"
+#include "tilt/pass/codegen/llvmgen.h"
+#include "tilt/pass/codegen/vinstr.h"
 #include "tilt/engine/engine.h"
 
 #include "test_base.h"
