+++ tilt/include/tilt/pass/codegen/llvmgen.h
@@ -6,7 +6,7 @@
 #include <string>
 #include <vector>
 
-#include "tilt/codegen/irgen.h"
+#include "tilt/pass/irgen.h"
 #include "llvm/IR/LLVMContext.h"
 #include "llvm/IR/IRBuilder.h"
 #include "llvm/IR/Module.h"
